Various fixes and cleanup
