+++ nautobot/docs/development/core/application-registry.md
@@ -197,9 +197,6 @@ List of GraphQL Type objects that will be added to the GraphQL schema. GraphQL o
 ]
 ```
 
---- 2.0.0
-    The `plugin_jobs` registry has been replaced by [`nautobot.core.celery.register_jobs`](../../user-guide/platform-functionality/jobs/index.md#writing-jobs) which should be called at import time by any App that provides jobs.
-
 ### `plugin_template_extensions`
 
 App content that gets embedded into core Nautobot templates. The store comprises Nautobot models registered as dictionary keys, each pointing to a list of applicable template extension classes that exist. An example:
