+++ nautobot/docs/development/jobs/migration/from-v1.md
@@ -111,8 +111,10 @@ The `read_only` Job field no longer forces an automatic database rollback at the
 ### Job Registration
 
 +/- 2.0.0
+    All Job subclasses, including `JobHookReceiver` and `JobButtonReceiver` subclasses must be registered at **import time** using the `nautobot.apps.jobs.register_jobs` method. This method accepts one or more job classes as arguments. You must account for how your jobs are imported when deciding where to call this method.
 
-All Job classes, including `JobHookReceiver` and `JobButtonReceiver` classes must be registered at **import time** using the `nautobot.apps.jobs.register_jobs` method. This method accepts one or more job classes as arguments. You must account for how your jobs are imported when deciding where to call this method.
++/- 2.2.2
+    Removed the requirement to call `register_jobs()` on module import. Any subclasses of `Job` that are defined at import time will automatically be registered.
 
 #### Registering Jobs in JOBS_ROOT or Git Repositories
 
@@ -123,12 +125,10 @@ from . import my_job_module
 ```
 
 ```py title="$JOBS_ROOT/my_jobs/my_job_module.py"
-from nautobot.apps.jobs import Job, register_jobs
+from nautobot.apps.jobs import Job
 
 class MyJob(Job):
     ...
-
-register_jobs(MyJob)
 ```
 
 Similarly, only the `jobs` module is loaded from Git repositories. If you're using submodules, you need to ensure that your jobs are either registered in the repository's `jobs/__init__.py` or that this file imports your submodules where the jobs are registered.
