+++ nautobot/extras/test_jobs/api_test_job.py
@@ -1,6 +1,5 @@
 from django.conf import settings
 
-from nautobot.core.celery import register_jobs
 from nautobot.extras.jobs import BooleanVar, get_task_logger, IntegerVar, Job, ObjectVar, StringVar
 from nautobot.extras.models import Role
 
@@ -25,6 +24,3 @@ class APITestJob(Job):
         logger.error(var4)
 
         return "Job complete"
-
-
-register_jobs(APITestJob)
