+++ nautobot/extras/test_jobs/dry_run.py
@@ -1,4 +1,3 @@
-from nautobot.core.celery import register_jobs
 from nautobot.extras.jobs import DryRunVar, get_task_logger, Job
 from nautobot.extras.models import Status
 
@@ -22,6 +21,3 @@ class TestDryRun(Job):
         if not dryrun:
             obj.save()
         logger.info("Status created successfully.", extra={"object": obj})
-
-
-register_jobs(TestDryRun)
