+++ nautobot/extras/test_jobs/duplicate_name.py
@@ -1,4 +1,3 @@
-from nautobot.core.celery import register_jobs
 from nautobot.extras.jobs import Job
 
 
@@ -22,6 +21,3 @@ class TestDuplicateName2(Job):
 
 class TestDuplicateNameNoMeta(Job):
     pass
-
-
-register_jobs(TestDuplicateName1, TestDuplicateName2, TestDuplicateNameNoMeta)
