+++ nautobot/extras/test_jobs/duplicate_name2.py
@@ -1,9 +1,5 @@
-from nautobot.core.celery import register_jobs
 from nautobot.extras.jobs import Job
 
 
 class TestDuplicateNameNoMeta(Job):
     pass
-
-
-register_jobs(TestDuplicateNameNoMeta)
