+++ nautobot/extras/test_jobs/fail.py
@@ -1,4 +1,3 @@
-from nautobot.core.celery import register_jobs
 from nautobot.extras.jobs import get_task_logger, Job, RunJobTaskFailed
 
 logger = get_task_logger(__name__)
@@ -46,6 +45,3 @@ class TestFailWithSanitization(Job):
             b"Cloning into '/Users/jathan/.nautobot/git/git_test'...\nfatal: could not read Password for https://abc123@github.com': terminal prompts disabled\n",
         )
         raise exc
-
-
-register_jobs(TestFail, TestFailWithSanitization)
