+++ nautobot/extras/test_jobs/field_default.py
@@ -1,4 +1,3 @@
-from nautobot.core.celery import register_jobs
 from nautobot.extras.jobs import IntegerVar, Job
 
 
@@ -14,6 +13,3 @@ class TestFieldDefault(Job):
         """Metaclass attrs."""
 
         field_order = ["var_int", "var_int_no_default"]
-
-
-register_jobs(TestFieldDefault)
