+++ nautobot/extras/test_jobs/field_order.py
@@ -1,4 +1,3 @@
-from nautobot.core.celery import register_jobs
 from nautobot.extras.jobs import FileVar, Job, StringVar
 
 
@@ -15,6 +14,3 @@ class TestFieldOrder(Job):
         """Metaclass attrs."""
 
         field_order = ["var1", "var2", "var23"]
-
-
-register_jobs(TestFieldOrder)
