+++ nautobot/extras/test_jobs/file_upload_fail.py
@@ -1,4 +1,3 @@
-from nautobot.core.celery import register_jobs
 from nautobot.extras.jobs import FileVar, get_task_logger, Job
 
 logger = get_task_logger(__name__)
@@ -26,6 +25,3 @@ class TestFileUploadFail(Job):
         logger.warning("File contents: %s", contents)
 
         raise self.exception("Test failure")
-
-
-register_jobs(TestFileUploadFail)
