+++ nautobot/extras/test_jobs/ipaddress_vars.py
@@ -1,6 +1,5 @@
 import json
 
-from nautobot.core.celery import register_jobs
 from nautobot.extras.jobs import get_task_logger, IPAddressVar, IPAddressWithMaskVar, Job
 
 logger = get_task_logger(__name__)
@@ -58,6 +57,3 @@ class TestIPAddresses(Job):
         logger.info("Job didn't crash!")
 
         return "Nice IPs, bro."
-
-
-register_jobs(TestIPAddresses)
