+++ nautobot/extras/test_jobs/log_redaction.py
@@ -1,4 +1,3 @@
-from nautobot.core.celery import register_jobs
 from nautobot.extras.jobs import get_task_logger, Job
 
 logger = get_task_logger(__name__)
@@ -14,6 +13,3 @@ class TestLogRedaction(Job):
         logger.warning("The secret is supersecret123")
         logger.error("The secret is supersecret123")
         logger.critical("The secret is supersecret123")
-
-
-register_jobs(TestLogRedaction)
