+++ nautobot/extras/test_jobs/log_skip_db_logging.py
@@ -1,4 +1,3 @@
-from nautobot.core.celery import register_jobs
 from nautobot.extras.jobs import get_task_logger, Job
 
 logger = get_task_logger(__name__)
@@ -11,6 +10,3 @@ class TestLogSkipDBLogging(Job):
     def run(self):
         logger.debug("I should NOT be logged to the database", extra={"skip_db_logging": True})
         logger.info("I should be logged to the database")
-
-
-register_jobs(TestLogSkipDBLogging)
