+++ nautobot/extras/test_jobs/modify_db.py
@@ -1,4 +1,3 @@
-from nautobot.core.celery import register_jobs
 from nautobot.extras.jobs import get_task_logger, Job
 from nautobot.extras.models import Status
 
@@ -19,6 +18,3 @@ class TestModifyDB(Job):
         )
         obj.save()
         logger.info("Status created successfully.", extra={"object": obj})
-
-
-register_jobs(TestModifyDB)
