+++ nautobot/extras/test_jobs/no_field_order.py
@@ -1,4 +1,3 @@
-from nautobot.core.celery import register_jobs
 from nautobot.extras.jobs import Job, StringVar
 
 
@@ -17,6 +16,3 @@ class BaseJobClassWithVariable(Job):
 class TestDefaultFieldOrderWithInheritance(BaseJobClassWithVariable):
     b_testvar2 = StringVar(description="This var should be second")
     a_testvar3 = StringVar(description="This var should be third")
-
-
-register_jobs(TestDefaultFieldOrderWithInheritance, TestNoFieldOrder)
