+++ nautobot/extras/test_jobs/object_var_required.py
@@ -1,4 +1,3 @@
-from nautobot.core.celery import register_jobs
 from nautobot.dcim.models import Location
 from nautobot.extras.jobs import get_task_logger, Job, ObjectVar
 
@@ -15,6 +14,3 @@ class TestRequiredObjectVar(Job):
     def run(self, location):
         logger.info("The Location that the user provided.", extra={"object": location})
         return "Nice Location!"
-
-
-register_jobs(TestRequiredObjectVar)
