+++ nautobot/extras/test_jobs/profiling.py
@@ -1,7 +1,6 @@
 import pstats
 import tempfile
 
-from nautobot.core.celery import register_jobs
 from nautobot.extras.jobs import get_task_logger, Job
 
 logger = get_task_logger(__name__)
@@ -26,6 +25,3 @@ class TestProfilingJob(Job):
     def after_return(self, status, retval, task_id, args, kwargs, einfo):
         super().after_return(status, retval, task_id, args, kwargs, einfo=einfo)
         pstats.Stats(f"{tempfile.gettempdir()}/nautobot-jobresult-{self.job_result.id}.pstats")
-
-
-register_jobs(TestProfilingJob)
