+++ nautobot/extras/test_jobs/required_args.py
@@ -1,9 +1,5 @@
-from nautobot.core.celery import register_jobs
 from nautobot.extras.jobs import Job, StringVar
 
 
 class TestRequired(Job):
     var = StringVar(required=True)
-
-
-register_jobs(TestRequired)
