+++ nautobot/extras/test_jobs/soft_time_limit_greater_than_time_limit.py
@@ -1,4 +1,3 @@
-from nautobot.core.celery import register_jobs
 from nautobot.extras.jobs import Job
 
 
@@ -10,6 +9,3 @@ class TestSoftTimeLimitGreaterThanHardTimeLimit(Job):
 
     def run(self):
         pass
-
-
-register_jobs(TestSoftTimeLimitGreaterThanHardTimeLimit)
