+++ nautobot/extras/test_jobs/task_queues.py
@@ -1,4 +1,3 @@
-from nautobot.core.celery import register_jobs
 from nautobot.extras.jobs import Job
 
 
@@ -18,6 +17,3 @@ class TestWorkerQueues(Job):
 
     def run(self):
         pass
-
-
-register_jobs(TestWorkerQueues)
